Fix:suprresed warnings and unused variables
